requirements.txt
Flask==2.0.1
Pillow==8.4.0
gTTS==2.2.3
moviepy==1.0.3
python-dotenv==0.19.0
werkzeug==2.0.1
gunicorn==20.1.0 